+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/ModuleResolvers/SchemaConfigurationFunctionalityModuleResolver.php
@@ -127,7 +127,7 @@ class SchemaConfigurationFunctionalityModuleResolver extends AbstractFunctionali
     public function getSettingsDefaultValue(string $module, string $option): mixed
     {
         // Lower the security constraints for the static app
-        $useUnsafe = BehaviorHelpers::areUnsafeDefaultsEnabled();
+        $useNonRestrictive = BehaviorHelpers::areNonRestrictiveDefaultsEnabled();
         $defaultValues = [
             self::SCHEMA_NAMESPACING => [
                 ModuleSettingOptions::DEFAULT_VALUE => false,
@@ -139,7 +139,7 @@ class SchemaConfigurationFunctionalityModuleResolver extends AbstractFunctionali
                 ModuleSettingOptions::DEFAULT_VALUE => MutationSchemes::STANDARD,
             ],
             self::SCHEMA_EXPOSE_SENSITIVE_DATA => [
-                ModuleSettingOptions::DEFAULT_VALUE => $useUnsafe,
+                ModuleSettingOptions::DEFAULT_VALUE => $useNonRestrictive,
             ],
             self::SCHEMA_SELF_FIELDS => [
                 ModuleSettingOptions::DEFAULT_VALUE => true,